fix bpf integration test
